+++ lib/huge_listview.dart
@@ -3,5 +3,6 @@ library huge_listview;
 export 'package:huge_listview/src/draggable_scrollbar.dart';
 export 'package:huge_listview/src/draggable_scrollbar_thumbs.dart';
 export 'package:huge_listview/src/huge_listview.dart';
+export 'package:huge_listview/src/huge_listview_controller.dart';
 export 'package:huge_listview/src/page_result.dart';
 export 'package:huge_listview/src/slidefade_transition.dart';