+++ lib/src/slidefade_transition.dart
@@ -4,13 +4,16 @@ class SlideFadeTransition extends StatelessWidget {
   final Animation<double> animation;
   final Widget child;
 
-  const SlideFadeTransition({Key? key, required this.animation, required this.child}) : super(key: key);
+  const SlideFadeTransition(
+      {Key? key, required this.animation, required this.child})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
       animation: animation,
-      builder: (context, child) => animation.value == 0.0 ? Container() : child!,
+      builder: (context, child) =>
+          animation.value == 0.0 ? Container() : child!,
       child: SlideTransition(
         position: Tween(
           begin: const Offset(0.3, 0.0),